Hogeクラスを静的ライブラリ化
