+++ cmake/src/main.cpp
@@ -1,50 +1,6 @@
 #include <iostream>
 
-/**
- *  @class  Hoge
- *  @brief  int型の値を扱うクラスです
- */
-class Hoge {
-private:
-    int private_num;
-
-public:
-    Hoge();
-    Hoge(int n);
-    void Display();
-    ~Hoge();
-};
-
-/**
- *  @fn     Hoge
- *  @brief  デフォルトコンストラクタ。Hogeオブジェクトを作成します
- */
-Hoge::Hoge(){
-    private_num = 0;
-}
-
-/**
- *  @fn     Hoge
- *  @brief  引数付きコンストラクタ。Hogeオブジェクトを作成します
- *  @param  初期化に指定する値
- */
-Hoge::Hoge(int n){
-    private_num = n;
-}
-
-/**
- *  @fn     Disp
- *  @brief  メンバ変数の表示
- */
-void Hoge::Display(){
-    std::cout << "num = " << private_num << std::endl;
-}
-
-/**
- *  @fn     ~Hoge
- *  @brief  デストラクタ
- */
-Hoge::~Hoge(){}
+#include "hoge.hpp"
 
 int main(){
     // 引数なし宣言の場合
